requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
dj-database-url==0.5.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-heroku==0.3.1
django-taggit==1.5.1
django==3.2.6
gunicorn==20.1.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1; python_version >= '3.5'
whitenoise==5.3.0
